coral-ai/requirements.txt
werkzeug==2.0.3
waitress==2.1.2
flask-cors==3.0.10
flask==2.1.1
requests==2.27.1
Pillow==9.0.1
keras==2.14.0
tensorflow==2.14.0
numpy==1.25.0
